pyproject.toml
[project]
name = "curvepack2d"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
  "beartype>=0.18.5",
  "numpy",
  "scipy",
  "matplotlib",
  "jax",
  "jaxlib",
  "optax",
  "svgpathtools",
  "shapely",
  "tqdm",
  "Pillow",
  "svgwrite",
  "jaxtyping>=0.3.7",
  "pyright>=1.1.408",
  "pytest>=9.0.2",
]

[project.optional-dependencies]
dev = [
  "pytest>=9.0.2",
]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["src", "src.*"]
